wrap iframes in a 16x9 padded div
in the content rendering function.  This allows for fluid control of videos; basically all of the iframes  in the content are 16x9 videos. This is maybe a bit fragile, but it appears to be effective.  Inspired by numerous articles on the subject including https://css-tricks.com/NetMag/FluidWidthVideo/Article-FluidWidthVideo.php